+++ package.json
@@ -9,6 +9,9 @@
     "web": "expo start --web"
   },
   "dependencies": {
+    "@react-native-community/masked-view": "^0.1.11",
+    "@react-navigation/native": "^6.1.6",
+    "@react-navigation/stack": "^6.3.16",
     "expo": "~48.0.15",
     "expo-location": "~15.1.1",
     "expo-status-bar": "~1.4.4",
@@ -16,6 +19,10 @@
     "react-native": "0.71.7",
     "react-native-chart-kit": "^6.12.0",
     "react-native-fs": "^2.20.0",
+    "react-native-gesture-handler": "~2.9.0",
+    "react-native-reanimated": "~2.14.4",
+    "react-native-safe-area-context": "4.5.0",
+    "react-native-screens": "~3.20.0",
     "react-native-svg": "^13.9.0",
     "react-native-xml2js": "^1.0.3",
     "xlsx": "^0.18.5",
