+++ screens/Test.js
@@ -1,0 +1,13 @@
+import React, {useEffect, useState} from "react";
+import { View, Text, Dimensions, StyleSheet, ScrollView, TextInput, Button } from 'react-native';
+
+function Test({navigation}){
+    return(
+        <View>
+            <Text>hi</Text>
+        </View>
+        
+    );
+}
+
+export default Test;
